+++ internal/timestamp/testdata/TimeStampTokenWithoutSigningCertificateV1orV2.p7s
