+++ revocation/revocation_test.go
@@ -469,7 +469,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 	}
 
 	t.Run("empty chain", func(t *testing.T) {
-		certResults, err := ValidateTimestampCertChain([]*x509.Certificate{}, time.Now(), &http.Client{Timeout: 5 * time.Second})
+		r, err := NewTimestamp(&http.Client{Timeout: 5 * time.Second})
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate([]*x509.Certificate{}, time.Now())
 		expectedErr := result.InvalidChainError{Err: errors.New("chain does not contain any certificates")}
 		if err == nil || err.Error() != expectedErr.Error() {
 			t.Errorf("Expected CheckStatus to fail with %v, but got: %v", expectedErr, err)
@@ -480,7 +484,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 	})
 	t.Run("check non-revoked chain", func(t *testing.T) {
 		client := testhelper.MockClient(testChain, []ocsp.ResponseStatus{ocsp.Good}, nil, true)
-		certResults, err := ValidateTimestampCertChain(revokableChain, time.Now(), client)
+		r, err := NewTimestamp(client)
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate(revokableChain, time.Now())
 		if err != nil {
 			t.Errorf("Expected CheckStatus to succeed, but got error: %v", err)
 		}
@@ -497,7 +505,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 	t.Run("check chain with 1 Unknown cert", func(t *testing.T) {
 		// 3rd cert will be unknown, the rest will be good
 		client := testhelper.MockClient(testChain, []ocsp.ResponseStatus{ocsp.Good, ocsp.Good, ocsp.Unknown, ocsp.Good}, nil, true)
-		certResults, err := ValidateTimestampCertChain(revokableChain, time.Now(), client)
+		r, err := NewTimestamp(client)
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate(revokableChain, time.Now())
 		if err != nil {
 			t.Errorf("Expected CheckStatus to succeed, but got error: %v", err)
 		}
@@ -519,7 +531,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 	t.Run("check OCSP with 1 revoked cert", func(t *testing.T) {
 		// 3rd cert will be revoked, the rest will be good
 		client := testhelper.MockClient(testChain, []ocsp.ResponseStatus{ocsp.Good, ocsp.Good, ocsp.Revoked, ocsp.Good}, nil, true)
-		certResults, err := ValidateTimestampCertChain(revokableChain, time.Now(), client)
+		r, err := NewTimestamp(client)
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate(revokableChain, time.Now())
 		if err != nil {
 			t.Errorf("Expected CheckStatus to succeed, but got error: %v", err)
 		}
@@ -541,7 +557,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 	t.Run("check OCSP with 1 unknown and 1 revoked cert", func(t *testing.T) {
 		// 3rd cert will be unknown, 5th will be revoked, the rest will be good
 		client := testhelper.MockClient(testChain, []ocsp.ResponseStatus{ocsp.Good, ocsp.Good, ocsp.Unknown, ocsp.Good, ocsp.Revoked, ocsp.Good}, nil, true)
-		certResults, err := ValidateTimestampCertChain(revokableChain, time.Now(), client)
+		r, err := NewTimestamp(client)
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate(revokableChain, time.Now())
 		if err != nil {
 			t.Errorf("Expected CheckStatus to succeed, but got error: %v", err)
 		}
@@ -569,7 +589,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 		revokedTime := time.Now().Add(time.Hour)
 		// 3rd cert will be future revoked, the rest will be good
 		client := testhelper.MockClient(testChain, []ocsp.ResponseStatus{ocsp.Good, ocsp.Good, ocsp.Revoked, ocsp.Good}, &revokedTime, true)
-		certResults, err := ValidateTimestampCertChain(revokableChain, time.Now(), client)
+		r, err := NewTimestamp(client)
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate(revokableChain, time.Now())
 		if err != nil {
 			t.Errorf("Expected CheckStatus to succeed, but got error: %v", err)
 		}
@@ -587,7 +611,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 		revokedTime := time.Now().Add(time.Hour)
 		// 3rd cert will be unknown, 5th will be future revoked, the rest will be good
 		client := testhelper.MockClient(testChain, []ocsp.ResponseStatus{ocsp.Good, ocsp.Good, ocsp.Unknown, ocsp.Good, ocsp.Revoked, ocsp.Good}, &revokedTime, true)
-		certResults, err := ValidateTimestampCertChain(revokableChain, time.Now(), client)
+		r, err := NewTimestamp(client)
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate(revokableChain, time.Now())
 		if err != nil {
 			t.Errorf("Expected CheckStatus to succeed, but got error: %v", err)
 		}
@@ -609,7 +637,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 	t.Run("check OCSP with 1 revoked cert before signing time", func(t *testing.T) {
 		// 3rd cert will be revoked, the rest will be good
 		client := testhelper.MockClient(testChain, []ocsp.ResponseStatus{ocsp.Good, ocsp.Good, ocsp.Revoked, ocsp.Good}, nil, true)
-		certResults, err := ValidateTimestampCertChain(revokableChain, time.Now().Add(time.Hour), client)
+		r, err := NewTimestamp(client)
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate(revokableChain, time.Now().Add(time.Hour))
 		if err != nil {
 			t.Errorf("Expected CheckStatus to succeed, but got error: %v", err)
 		}
@@ -635,7 +667,11 @@ func TestCheckRevocationStatusForTimestampChain(t *testing.T) {
 		if !zeroTime.IsZero() {
 			t.Errorf("exected zeroTime.IsZero() to be true")
 		}
-		certResults, err := ValidateTimestampCertChain(revokableChain, time.Now().Add(time.Hour), client)
+		r, err := NewTimestamp(client)
+		if err != nil {
+			t.Errorf("Expected successful creation of revocation, but received error: %v", err)
+		}
+		certResults, err := r.Validate(revokableChain, time.Now().Add(time.Hour))
 		if err != nil {
 			t.Errorf("Expected CheckStatus to succeed, but got error: %v", err)
 		}
